fix(bing): spotbug issue
Former-commit-id: d8aab5284cc498805493bba864876490e1cc23ba [formerly 6590872edad6c9afdebda56b7a57f0be0e7af700]
Former-commit-id: c10ab1e0c787d1d8ebed40893a999c8064497637

+++ connectors/BingConnector/src/de/l3s/interwebj/connector/bing/BingConnector.java
@@ -326,7 +326,7 @@ public class BingConnector extends AbstractServiceConnector implements Cloneable
                     .withQuery(query.getQuery())
                     .withMarket(language)
                     .withCount(count)
-                    .withOffset((long) ((query.getPage() - 1) * count))
+                    .withOffset(((query.getPage() - 1L) * count))
                     .execute();
 
             // Results go here
